+++ packages/analytics/src/index.js
@@ -30,6 +30,11 @@ import multiply from './mulitply';
 import distance from './distance';
 import aggregate from './aggregate';
 import statistics from './statistics';
+import combine from './combine';
+import expand from './expand';
+import files from './files';
+import bufferize from './bufferize';
+import buffers from './buffers';
 
 export default {
     count,
@@ -64,4 +69,9 @@ export default {
     distance,
     aggregate,
     statistics,
+    combine,
+    expand,
+    files,
+    bufferize,
+    buffers,
 };