+++ packages/analytics/test/tests.js
@@ -1,13 +1,15 @@
-const assert = require('assert');
-const from = require('from');
-const ezs = require('../../core/src');
+import assert from 'assert';
+import from from 'from';
+import ezs from '../../core/src';
+import statements from '../src';
+
+ezs.addPath(__dirname);
 
 const data01 = require('./data01.json');
 
-ezs.use(require('../src'));
-
 describe('test', () => {
     it('distinct', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 'z' },
@@ -31,6 +33,7 @@ describe('test', () => {
             });
     });
     it('reducing', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 'z' },
@@ -53,6 +56,7 @@ describe('test', () => {
             });
     });
     it('summing', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 'z' },
@@ -76,6 +80,7 @@ describe('test', () => {
             });
     });
     it('combining', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 'x', value: 2 },
@@ -96,7 +101,8 @@ describe('test', () => {
                 done();
             });
     });
-    it('count', (done) => {
+    it('count #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 'z' },
@@ -118,7 +124,32 @@ describe('test', () => {
                 done();
             });
     });
-    it('pluck', (done) => {
+    it('count #2', (done) => {
+        ezs.use(statements);
+        const res = [];
+        from([
+            { a: 'x', b: 'z' },
+            { a: 't', b: 'z' },
+            { a: 't', b: 'z' },
+            { c: 'x', b: 'z' },
+            { a: 'x', b: 'z' },
+        ])
+            .pipe(ezs('count', { path: ['a', 'b'] }))
+            .pipe(ezs('reducing'))
+            .pipe(ezs('summing'))
+            .on('data', (chunk) => {
+                assert(typeof chunk === 'object');
+                res.push(chunk);
+            })
+            .on('end', () => {
+                assert.equal(2, res.length);
+                assert.equal(4, res[0].value);
+                assert.equal(5, res[1].value);
+                done();
+            });
+    });
+    it('pluck #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: ['m', 'n', 'o'] },
@@ -140,7 +171,29 @@ describe('test', () => {
                 done();
             });
     });
+    it('pluck #2', (done) => {
+        ezs.use(statements);
+        const res = [];
+        from([
+            { a: 'x', b: 'z' },
+            { a: 't', b: 'z' },
+            { a: 't', b: 'z' },
+            { c: 'x', b: 'z' },
+            { a: 'x', b: 'z' },
+        ])
+            .pipe(ezs('pluck', { path: 'b' }))
+            .on('data', (chunk) => {
+                assert(typeof chunk === 'object');
+                assert.equal('z', chunk.value);
+                res.push(chunk);
+            })
+            .on('end', () => {
+                assert.equal(5, res.length);
+                done();
+            });
+    });
     it('keys', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 1, b: 2 },
@@ -165,6 +218,7 @@ describe('test', () => {
     });
 
     it('maximizing', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 1, value: [2, 3, 4, 5, 1] },
@@ -187,6 +241,7 @@ describe('test', () => {
     });
 
     it('minimizing', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 1, value: [2, 3, 4, 5, 1] },
@@ -209,6 +264,7 @@ describe('test', () => {
     });
 
     it('merging #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 1, value: { a: 1 } },
@@ -237,6 +293,7 @@ describe('test', () => {
     });
 
     it('merging #2', (done) => {
+        ezs.use(statements);
         const script = `
             [replace]
             path = id
@@ -292,6 +349,7 @@ describe('test', () => {
 
 
     it('totalize', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: '1', b: '4' },
@@ -316,6 +374,7 @@ describe('test', () => {
     });
 
     it.skip('exploding', (done) => {
+        ezs.use(statements);
         const script = `
             [exploding]
             id = id
@@ -368,6 +427,7 @@ describe('test', () => {
     });
 
     it('split & count', (done) => {
+        ezs.use(statements);
         const script = `
             [replace]
             path = id
@@ -412,6 +472,7 @@ describe('test', () => {
     });
 
     it('slice #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22])
             .pipe(ezs('slice'))
@@ -427,6 +488,7 @@ describe('test', () => {
     });
 
     it('slice #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22])
             .pipe(ezs('slice', { start: 2, size: 2 }))
@@ -442,6 +504,7 @@ describe('test', () => {
     });
 
     it('slice #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
             .pipe(ezs('slice'))
@@ -457,6 +520,7 @@ describe('test', () => {
     });
 
     it('distribute #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 1, b: 'a' },
@@ -492,6 +556,7 @@ describe('test', () => {
     });
 
     it('distribute #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 1, b: 'a' },
@@ -526,6 +591,7 @@ describe('test', () => {
     });
 
     it('distribute #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: '2013', value: 8 },
@@ -554,6 +620,7 @@ describe('test', () => {
     });
 
     it('distribute #4', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: '2013', value: 8 },
@@ -581,6 +648,7 @@ describe('test', () => {
     });
 
     it('distribute #5', (done) => {
+        ezs.use(statements);
         const res = [];
         from(data01)
             .pipe(ezs('groupingByModulo', { modulo: 1 }))
@@ -596,6 +664,7 @@ describe('test', () => {
     });
 
     it('greater #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -620,6 +689,7 @@ describe('test', () => {
     });
 
     it('greater #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -644,6 +714,7 @@ describe('test', () => {
     });
 
     it('greater #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -668,6 +739,7 @@ describe('test', () => {
     });
 
     it('less #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -692,6 +764,7 @@ describe('test', () => {
     });
 
     it('less #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -717,6 +790,7 @@ describe('test', () => {
 
 
     it('less #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -741,6 +815,7 @@ describe('test', () => {
     });
 
     it('drop #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -763,6 +838,7 @@ describe('test', () => {
     });
 
     it('drop #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -785,6 +861,7 @@ describe('test', () => {
     });
 
     it('drop #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -807,6 +884,7 @@ describe('test', () => {
     });
 
     it('drop #4', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -830,6 +908,7 @@ describe('test', () => {
 
 
     it('drop #5', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -852,6 +931,7 @@ describe('test', () => {
     });
 
     it('drop #6', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -874,6 +954,7 @@ describe('test', () => {
     });
 
     it('filter #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -897,6 +978,7 @@ describe('test', () => {
     });
 
     it('filter #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -920,6 +1002,7 @@ describe('test', () => {
     });
 
     it('filter #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -942,6 +1025,7 @@ describe('test', () => {
     });
 
     it('filter #4', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -964,6 +1048,7 @@ describe('test', () => {
     });
 
     it('filter #5', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { id: 2000, value: 1 },
@@ -990,6 +1075,7 @@ describe('test', () => {
     });
 
     it('multiply #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 2000, b: 1 },
@@ -1019,6 +1105,7 @@ describe('test', () => {
     });
 
     it('multiply #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 2000, b: 1 },
@@ -1040,6 +1127,7 @@ describe('test', () => {
     });
 
     it('multiply #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 2000, b: 1 },
@@ -1061,6 +1149,7 @@ describe('test', () => {
     });
 
     it('multiply #4', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 2000, b: 1 },
@@ -1082,6 +1171,7 @@ describe('test', () => {
     });
 
     it('multiply #5', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 2000, b: 1 },
@@ -1102,6 +1192,7 @@ describe('test', () => {
             });
     });
     it('aggregate #1', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 'z' },
@@ -1124,6 +1215,7 @@ describe('test', () => {
             });
     });
     it('aggregate #2', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: ['x', 'x'], b: 'z' },
@@ -1146,6 +1238,7 @@ describe('test', () => {
             });
     });
     it('aggregate #3', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 3 },
@@ -1169,6 +1262,7 @@ describe('test', () => {
     });
 
     it('aggregate #4', (done) => {
+        ezs.use(statements);
         const res = [];
         from([
             { a: 'x', b: 3 },
@@ -1188,7 +1282,4 @@ describe('test', () => {
                 done();
             });
     });
-
-
-
 });