+++ packages/storage/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/Inist-CNRS/ezs/compare/@ezs/storage@1.2.0...@ezs/storage@1.2.1) (2020-06-12)
+
+**Note:** Version bump only for package @ezs/storage
+
+
+
+
+
 # [1.2.0](https://github.com/Inist-CNRS/ezs/compare/@ezs/storage@1.1.10...@ezs/storage@1.2.0) (2020-05-11)